Fix #297 -- Add Django admin' inline formset support (#300)
As described in #249 and #297 django-select2 does not currently work
with selects added dynamically in the Django admin.

The solution was implemented in #249 and all credit should go to Jurrian
Tromp. I only added a small correction:

```diff
- jqRow.find('.select2-container').remove()
+ jqRow.find('.django-select2').parent().find('.select2-container').remove()
```

to only remove `.select2-container` when it is a sibling of
`.django-select2`. Otherwise the wrong `.select2-container` might get
deleted.

---------

Co-authored-by: Johannes Maron <johannes@maron.family>